Miroslav Trmac writes: The 'port' setting used to be an integer, now it can be an integer or a string, depending on the presence of IPv6 support.  fetchmailconf assumes 'port' is an integer, though.
When 'port' is a string, it defaults to None; fetchmailconf can't
handle that and writes 'port None' to the output config file if the
server entry is not edited, see
https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=55623
The attached patch fixes this.

Note that fetchmailconf still doesn't support other string values for
'port', even with this patch.
    Mirek

svn path=/trunk/; revision=4058