Berlios Patch #2525, courtesy of Diego E. "Flameeyes" Pettenò:
"The AC_CANONICAL_TARGET macro and the $target_os variables are used for the
target of compilers and other code-generation tools, and should not be used
during cross-compile of generic software. Replace them with
AC_CANONICAL_HOST and $host_os instead, as they should have been from the
start.

For a breakdown of what host, build and target machines are, please see
http://blog.flameeyes.eu/s/canonical-target ."

svn path=/branches/BRANCH_6-3/; revision=5258