Major progress ticker bugfix/overhaul.
Progress tickers had been used inconsistently for a long time, and
documentation was outdated, too. Factor out common code to ease
maintenance, use the report_flush() function, and add and use a macro
(want_progress()) to determine if progress ticker output is desired.

This makes for a much more consistent look on screen and in logfiles and
should be much easier to fix later on.

TODO: test syslog output.

svn path=/branches/BRANCH_6-3/; revision=5290