Fix mimedecode last-line omission.
The mimedecode feature failed to ship the last line of the body if it
was encoded as quoted-printable and had a MIME soft line break in the
very last line.  Reported by Lars Hecking in June 2011.

Bug introduced on 1998-03-20 when the mimedecode support was added by
ESR before release 4.4.1 through code contributed by Henrik Storner,
in driver.c.

Workaround for older releases: do not use mimedecode feature.